create repository, service, controller for Category

+++ backend/src/main/java/com/netcracker/edu/backend/repository/CategoryEntityRepository.java
@@ -3,5 +3,5 @@ package com.netcracker.edu.backend.repository;
 import com.netcracker.edu.backend.entity.CategoryEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CategoryEntityRepository extends CrudRepository<CategoryEntity, Long> {
+public interface CategoryEntityRepository extends CrudRepository<CategoryEntity, Integer> {
 }
